requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
bcrypt==4.2.0
Django==4.2.15
django-cors==0.1
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
mysqlclient==2.2.4
packaging==24.1
PyJWT==2.9.0
pytz==2024.1
PyYAML==6.0.2
sqlparse==0.5.1
typing-extensions==4.12.2
uritemplate==4.1.1
